refactor: extract Project struct into its own file
